requirements.txt
numpy==1.24.1
pandas==1.5.3
Pillow==9.5.0
requests==2.28.2
streamlit==1.20.0
streamlit_extras==0.2.7
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.2
tensorflow==2.12.0
tensorflow_intel==2.11.0
